moving as much as possible to self-contained folder.
