+++ components/homme/src/share/partmcsl/CMakeLists.txt
@@ -1,8 +1,11 @@
 include_directories(
+  ${CMAKE_CURRENT_SOURCE_DIR}
   ${CMAKE_CURRENT_BINARY_DIR}
   ${HOMME_SOURCE_DIR}/src/share # for physgrid stuff (gllfvremap*mod.F90)
   ${HOMME_SOURCE_DIR}/src/share/compose # for SL infrastructure, polygonal intersection, and area quadrature
 )
+
+add_definitions(-DHOMME_ENABLE_PARTMCSL)
 
 message(STATUS "PARTMCSL stochastic transport enabled")
